api/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
python-multipart==0.0.6
PyMuPDF==1.23.8
pikepdf==9.11.0
Pillow==10.1.0
python-magic==0.4.27
aiofiles==23.2.1
pydantic-settings==2.1.0